Lazyload MUI, no message parser, split modules
