+++ src/components/molecules/login.tsx
@@ -6,7 +6,7 @@ import { useQueryClient } from 'react-query';
 import SocialLogin from '@components/molecules/socialLogin';
 import Modal from '@components/organisms/modal';
 import AndOnboarding from '@components/templates/andOnboarding';
-import { login } from '@lib/api';
+import { login } from '@lib/api/login';
 import { useLoginForgotPasswordMutation } from '@lib/generated/graphql';
 import { makeRegisterRoute } from '@lib/routes';
 import useLanguageRoute from '@lib/useLanguageRoute';
