+++ src/components/molecules/recordingButtonFavorite.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { useIsRecordingFavorited } from '@lib/api';
+import { useIsRecordingFavorited } from '@lib/api/useIsRecordingFavorited';
 import { BaseColors } from '@lib/constants';
 import { Scalars } from '@lib/generated/graphql';
