+++ src/components/molecules/teaseRecording.tsx
@@ -8,7 +8,7 @@ import Heading2 from '@components/atoms/heading2';
 import Heading3 from '@components/atoms/heading3';
 import Heading6 from '@components/atoms/heading6';
 import ProgressBar from '@components/atoms/progressBar';
-import { useIsRecordingFavorited } from '@lib/api';
+import { useIsRecordingFavorited } from '@lib/api/useIsRecordingFavorited';
 import { BaseColors } from '@lib/constants';
 import {
 	AndMiniplayerFragment,
