+++ src/components/templates/andMiniplayer.tsx
@@ -1,10 +1,11 @@
 import clsx from 'clsx';
+import dynamic from 'next/dynamic';
 import React, { PropsWithChildren, useContext } from 'react';
-
-import Miniplayer from '@components/organisms/miniplayer';
 
 import styles from './andMiniplayer.module.scss';
 import { PlaybackContext } from './andPlaybackContext';
+
+const LazyMiniplayer = dynamic(() => import('../organisms/miniplayer'));
 
 export default function AndMiniplayer({
 	children,
@@ -52,7 +53,7 @@ export default function AndMiniplayer({
 			>
 				{children}
 			</div>
-			{recording && <Miniplayer />}
+			{recording && <LazyMiniplayer />}
 		</>
 	);
 }
