+++ src/pages/[language]/account/logout.tsx
@@ -2,7 +2,7 @@ import { useRouter } from 'next/router';
 import React, { useEffect, useState } from 'react';
 import { FormattedMessage } from 'react-intl';
 
-import { useLogout } from '@lib/api';
+import { useLogout } from '@lib/api/useLogout';
 import { makeDiscoverRoute } from '@lib/routes';
 import useLanguageRoute from '@lib/useLanguageRoute';
