+++ src/pages/[language]/account/preferences.ts
@@ -1,7 +1,7 @@
 import { GetServerSidePropsContext } from 'next';
 
 import Preferences from '@containers/account/preferences';
-import { storeRequest } from '@lib/api';
+import { storeRequest } from '@lib/api/storeRequest';
 import { getAccountPreferencesData } from '@lib/generated/graphql';
 import getDehydratedProps, { DehydratedProps } from '@lib/getDehydratedProps';
