+++ src/pages/[language]/library/history.tsx
@@ -3,7 +3,7 @@ import { GetServerSidePropsContext, GetServerSidePropsResult } from 'next';
 import LibraryHistory, {
 	ILibraryHistoryProps,
 } from '@containers/library/history';
-import { storeRequest } from '@lib/api';
+import { storeRequest } from '@lib/api/storeRequest';
 import { getLanguageIdByRoute } from '@lib/getLanguageIdByRoute';
 
 export default LibraryHistory;
