+++ src/pages/[language]/library/playlists.ts
@@ -3,7 +3,7 @@ import { GetServerSidePropsContext, GetServerSidePropsResult } from 'next';
 import LibraryPlaylists, {
 	ILibraryPlaylistsProps,
 } from '@containers/library/playlist/list';
-import { storeRequest } from '@lib/api';
+import { storeRequest } from '@lib/api/storeRequest';
 import { getLanguageIdByRoute } from '@lib/getLanguageIdByRoute';
 
 export default LibraryPlaylists;
