+++ src/pages/[language]/search/sponsors/page/[i].ts
@@ -3,7 +3,7 @@ import { GetServerSidePropsContext, GetServerSidePropsResult } from 'next';
 import SearchSponsors, {
 	SearchSponsorsProps,
 } from '@containers/search/sponsors';
-import { storeRequest } from '@lib/api';
+import { storeRequest } from '@lib/api/storeRequest';
 import { getSearchResultsSponsors } from '@lib/generated/graphql';
 import { getPaginatedStaticProps } from '@lib/getPaginatedStaticProps';
